fix: simply storage and incident model. refactor polling
The previous implementation of storage and services was redundant, the same method had to be added to multiple interfaces.

New implementation of repositories relies solely on a storage object (like Postgres) implementing a repository. Any processing that needs to be done to data before sending to client can still be done in the repository method.

Polling and frontend were refactored to reflect these changes, as well polling no longer relies directly on the Incident model or a DB connection.